+++ api-types/Cargo.toml
@@ -7,6 +7,7 @@ rust-version.workspace = true
 version = "0.1.0"
 
 [features]
+fw_cfg = []
 pvmemcontrol = []
 
 [dependencies]
