+++ vmm/src/vm_config.rs
@@ -5,10 +5,10 @@
 use std::collections::HashSet;
 use std::net::IpAddr;
 use std::path::{Path, PathBuf};
-#[cfg(feature = "fw_cfg")]
-use std::str::FromStr;
 use std::{fs, result};
 
+#[cfg(feature = "fw_cfg")]
+use api_types::FwCfgItemList;
 #[cfg(feature = "pvmemcontrol")]
 use api_types::PvmemcontrolConfig;
 use api_types::{
@@ -867,55 +867,6 @@ impl Default for FwCfgConfig {
             acpi_tables: true,
             items: None,
         }
-    }
-}
-
-#[cfg(feature = "fw_cfg")]
-#[derive(Clone, Debug, Deserialize, Eq, PartialEq, Serialize)]
-pub struct FwCfgItemList {
-    #[serde(default)]
-    pub item_list: Vec<FwCfgItem>,
-}
-
-#[cfg(feature = "fw_cfg")]
-#[serde_with::skip_serializing_none]
-#[derive(Clone, Debug, Deserialize, Eq, PartialEq, Serialize)]
-pub struct FwCfgItem {
-    #[serde(default)]
-    pub name: String,
-    #[serde(default)]
-    pub file: Option<PathBuf>,
-    #[serde(default)]
-    pub string: Option<String>,
-}
-
-#[cfg(feature = "fw_cfg")]
-pub enum FwCfgItemError {
-    InvalidValue(String),
-}
-
-#[cfg(feature = "fw_cfg")]
-impl FromStr for FwCfgItemList {
-    type Err = FwCfgItemError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let body = s
-            .trim()
-            .strip_prefix('[')
-            .and_then(|s| s.strip_suffix(']'))
-            .ok_or_else(|| FwCfgItemError::InvalidValue(s.to_string()))?;
-
-        let mut fw_cfg_items: Vec<FwCfgItem> = vec![];
-        let items: Vec<&str> = body.split(':').collect();
-        for item in items {
-            fw_cfg_items.push(
-                FwCfgItem::parse(item)
-                    .map_err(|_| FwCfgItemError::InvalidValue(item.to_string()))?,
-            );
-        }
-        Ok(FwCfgItemList {
-            item_list: fw_cfg_items,
-        })
     }
 }
